+++ CHANGELOG.md
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Debug logging for time taken to format each file.
 - `OptimisingLineFormatter` to reflow all kinds of logical lines.
   - Configuration option `wrap_column` controls the line length.
+  - Configuration option `begin_style` controls the location of `begin` in control flow statements.
 - Method `process_cursor` to the `LogicalLinesReconstructor` trait.
 - Trait `CursorMetadata`.
 - Cursor tracking, via the `--cursor` CLI option.